Support load segments/channels in parallel
